refactor: updated global webpack, linting and sass usage configurations
