+++ .babelrc
@@ -1,15 +1,5 @@
 {
-  "env": {
-    "development": {
-      "presets": ["next/babel"]
-    },
-    "production": {
-      "presets": ["next/babel"]
-    },
-    "test": {
-      "presets": [["next/babel", { "preset-env": { "modules": "commonjs" } }]]
-    }
-  },
+  "presets": ["next/babel"],
   "plugins": [
     ["babel-plugin-root-import", { "rootPathSuffix": "client" }],
     ["transform-class-properties"]
