+++ client/components/Button/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import styles from './styles.css';
+import styles from './styles.scss';
 
 /**
  * Button
