+++ client/components/Header/styles.scss
@@ -5,7 +5,7 @@
 ========================================================================== */
 
 header {
-  background-color: var(--color-amaranth);
+  background-color: $primary-red;
 }
 
 .url {
