+++ client/components/Radio/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import first from 'lodash/first';
 import classNames from 'classnames';
-import styles from './styles.css';
+import styles from './styles.scss';
 
 /**
  * Radio:
