--- client/shared/styles/custom-medias.css
@@ -1,30 +1,0 @@
-/* ==========================================================================
-  Custom media queries
-========================================================================== */
-
-/* Use it like so: @media (--layout-small) { ... } */
-
-/* Use these if you developing desktop first and want to target lower resolutions */
-
-@custom-media --layout-lte-xsmall (width <= 480px);
-@custom-media --layout-lte-small (width <= 768px);
-@custom-media --layout-lte-medium (width <= 1024px);
-@custom-media --layout-lte-large (width <= 1440px);
-@custom-media --layout-lte-xlarge (width <= 1920px);
-
-/* Use these if you are developing mobile first and want to target higher resolutions */
-
-@custom-media --layout-gte-small (width > 480px);
-@custom-media --layout-gte-medium (width > 768px);
-@custom-media --layout-gte-large (width > 1024px);
-@custom-media --layout-gte-xlarge (width > 1440px);
-@custom-media --layout-gte-xxlarge (width > 1920px);
-
-/* Use these to target a specific range */
-
-@custom-media --layout-xsmall (width <= 480px);
-@custom-media --layout-small (width > 480) AND width <= 768px);
-@custom-media --layout-medium (width > 768px) AND (width <= 1024px);
-@custom-media --layout-large (width > 1024px) AND (width <= 1440px);
-@custom-media --layout-xlarge (width > 1440px) AND (width <= 1920px);
-@custom-media --layout-xxlarge (width > 1920px);
