+++ client/shared/styles/globals.scss
@@ -1,0 +1,144 @@
+@import "variables";
+@import "grid";
+@import "mixins/typography";
+
+/* ==========================================================================
+  Document
+========================================================================== */
+
+/*
+1. Stretch <html> stretch to fill our screen height
+2. Make children of html (body) occupy at least 100% of the screen
+3. Viewport is scalable and occupies at least 320px (iPhone 5)
+*/
+
+html {
+  min-width: 320px;
+  height: 0;
+  min-height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+/* ==========================================================================
+  Body
+========================================================================== */
+
+/*
+1. Force scroll always to prevent scrollbars to appear/disappear based on the page contents
+2. Make sure that we occupy 100% of our parent and allow our child elements to do the same
+3. Needed for IE11 otherwise flex wouldn't grow vertically, see https://stackoverflow.com/a/42930574
+*/
+
+body {
+  overflow-x: hidden;
+  overflow-y: scroll;
+  padding-bottom: $body-padding-bottom-xs;
+  display: flex;
+  flex: 1 0 auto;
+  flex-direction: column;
+  background-color: $primary-white;
+  color: $primary-black;
+  font-size: $browser-context;
+  font-weight: 400;
+  font-style: normal;
+  -moz-osx-font-smoothing: grayscale;
+  text-rendering: optimizeLegibility;
+  overflow-wrap: break-word; /* Break long words by default */
+
+  //Here, it should be included the base mixin font of the website,
+  //eg.: @include helvetica;
+
+  @include media("md") {
+    padding-bottom: $body-padding-bottom-md;
+  }
+}
+
+/* ==========================================================================
+  Anchors
+========================================================================== */
+
+a,
+a:focus,
+a:visited,
+a:hover,
+a:active {
+  outline: none !important;
+  color: $primary-white;
+  text-decoration: none !important;
+  transition: color $default-timing;
+}
+
+a[href^="tel"] {
+  color: inherit;
+}
+
+/* ==========================================================================
+  Headings
+========================================================================== */
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  margin-top: 0;
+  margin-bottom: 0;
+  font-style: normal;
+  line-height: 1.2em;
+}
+
+input {
+  border-radius: 0;
+}
+
+select::-ms-expand {
+  display: none;
+}
+
+select:focus option {
+  background-color: $primary-gray;
+  color: $primary-white;
+}
+
+input:focus,
+select:focus,
+textarea:focus,
+button:focus {
+  outline: none;
+}
+
+input:not([value=""]) {
+  border-color: $primary-white;
+}
+
+input:hover {
+  border-color: $primary-white;
+}
+
+@keyframes autofill {
+  to {
+    background: transparent;
+    color: #666;
+  }
+}
+
+input:-webkit-autofill {
+  -webkit-animation-name: autofill;
+  -webkit-animation-fill-mode: both;
+  -webkit-text-fill-color: $primary-white;
+}
+
+::placeholder {
+  /* Most modern browsers support this now. */
+  color: $secondary-gray;
+}
+
+/* ==========================================================================
+  Images
+========================================================================== */
+
+svg {
+  cursor: pointer;
+}
