--- client/shared/styles/index.css
@@ -1,9 +1,0 @@
-@import "_reboot";
-@import "layout";
-@import "typography";
-
-/*
-  Grids available:
-  - Bootstrap v3.3.7
-    @import "bootstrap-grid";
-*/
