--- client/shared/styles/layout.css
@@ -1,40 +1,0 @@
-/* ==========================================================================
-  Document
-========================================================================== */
-
-/*
-1. Stretch <html> stretch to fill our screen height
-2. Make children of html (body) occupy at least 100% of the screen
-3. Viewport is scalable and occupies at least 320px (iPhone 5)
-*/
-
-html {
-  min-width: 320px;
-  height: 0;
-  min-height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-/* ==========================================================================
-  Body & #root
-========================================================================== */
-
-/*
-1. Force scroll always to prevent scrollbars to appear/disappear based on the page contents
-2. Make sure that we occupy 100% of our parent and allow our child elements to do the same
-3. Needed for IE11 otherwise flex wouldn't grow vertically, see https://stackoverflow.com/a/42930574
-*/
-
-body {
-  overflow-y: scroll;
-  display: flex;
-  flex: 1 0 auto;
-  flex-direction: column;
-}
-
-:global(#root) {
-  display: flex;
-  flex: 1 0 auto;
-  flex-direction: column;
-}
