+++ client/shared/styles/mixins/_position.scss
@@ -1,0 +1,6 @@
+@mixin alignCenter() {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+}
