+++ client/shared/styles/mixins/_typography.scss
@@ -1,0 +1,58 @@
+/* ==========================================================================
+  Smooth & sharp
+========================================================================== */
+
+@mixin typographySmooth() {
+  -moz-osx-font-smoothing: grayscale; /* Font smoothing tweaks for MacOSx in Gecko */
+  -webkit-font-smoothing: antialiased; /* Font smoothing tweaks for MacOSx in Webkit */
+}
+
+@mixin typographySharp() {
+  -moz-osx-font-smoothing: initial; /* Turn off font smoothing tweaks for MacOSx in Gecko */
+  -webkit-font-smoothing: initial; /* Turn off font smoothing tweaks for MacOSx in Webkit */
+}
+
+/* ==========================================================================
+  Ellipsis
+========================================================================== */
+
+/* The container needs to have a width for it to work */
+
+@mixin typographyEllipsis() {
+  overflow: hidden; /* "overflow" value must be different from "visible" */
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow-wrap: normal;
+}
+
+@mixin typographyEllipsisMultiline($line-count, $max-height) {
+  max-height: $max-height;
+  overflow: hidden;
+
+  /* $line-count lines of text */
+  /* stylelint-disable */
+  display: -webkit-box;
+  text-overflow: ellipsis;
+  -webkit-line-clamp: $line-count;
+  -webkit-box-orient: vertical;
+  /* stylelint-enable */
+}
+
+/* ==========================================================================
+  Truncate
+========================================================================== */
+
+@mixin truncate($width) {
+  width: $width;
+  max-width: $width;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+@mixin notTruncated {
+  width: 100%;
+  overflow: visible;
+  text-overflow: inherit;
+  white-space: normal;
+}
