--- client/shared/styles/mixins/clearfix.css
@@ -1,11 +1,0 @@
-/* ==========================================================================
-  Utility mixin to clear children floats
-========================================================================== */
-
-@define-mixin clearfix $selector {
-  $selector::after {
-    content: "";
-    display: table;
-    clear: both;
-  }
-}
