--- client/shared/styles/typography.css
@@ -1,90 +1,0 @@
-@import "variables";
-@import "mixins/typography";
-
-/* ==========================================================================
-  Fonts
-
-  If you are going to include custom fonts please use the /shared/media/fonts
-  directory.
-========================================================================== */
-
-/* @font-face {
-  font-family: "CustomFont";
-  font-weight: normal;
-  font-style: normal;
-  src: url("../media/fonts/customfont/CustomFont.eot");
-  src:
-    url("../media/fonts/custom-font/CustomFont.otf") format("opentype"),
-    url("../media/fonts/custom-font/CustomFont.eot?#iefix") format("embedded-opentype"),
-    url("../media/fonts/custom-font/CustomFont.woff2") format("woff2"),
-    url("../media/fonts/custom-font/CustomFont.woff") format("woff"),
-    url("../media/fonts/custom-font/CustomFont.ttf") format("truetype");
-} */
-
-/* ==========================================================================
-  Typography Defaults
-========================================================================== */
-
-body {
-  color: var(--color-black);
-  font-family: var(--typography-text-font);
-  font-size: 16px;
-  overflow-wrap: break-word; /* Break long words by default */
-}
-
-strong {
-  font-weight: 500;
-}
-
-/* ==========================================================================
-  Links
-========================================================================== */
-
-a {
-  color: var(--color-amaranth);
-  transition: color 0.2s;
-}
-
-a:hover {
-  color: var(--color-amaranth);
-  text-decoration: underline;
-}
-
-/* ==========================================================================
-  Headings
-========================================================================== */
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: var(--typography-display-font);
-  font-weight: 400;
-  line-height: 1.1;
-}
-
-h1 {
-  font-size: 42px;
-}
-
-h2 {
-  font-size: 36px;
-}
-
-h3 {
-  font-size: 32px;
-}
-
-h4 {
-  font-size: 28px;
-}
-
-h5 {
-  font-size: 24px;
-}
-
-h6 {
-  font-size: 20px;
-}
