+++ package.json
@@ -79,6 +79,7 @@
     "@commitlint/cli": "^6.2.0",
     "@commitlint/config-conventional": "^6.1.0",
     "@zeit/next-css": "^0.2.0",
+    "@zeit/next-sass": "^0.2.0",
     "babel-eslint": "^8.2.6",
     "babel-plugin-root-import": "^5.1.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -92,6 +93,7 @@
     "husky": "^0.14.3",
     "lint-staged": "^7.2.2",
     "nock": "^9.6.1",
+    "node-sass": "^4.9.3",
     "postcss": "^6.0.23",
     "postcss-conditionals": "^2.1.0",
     "postcss-css-variables": "^0.8.1",
@@ -105,7 +107,9 @@
     "stylelint-config-standard": "^18.2.0",
     "stylelint-no-unsupported-browser-features": "^3.0.1",
     "stylelint-order": "^0.8.1",
+    "stylelint-scss": "^3.3.0",
     "svg-css-modules-loader": "^1.6.1",
+    "svg-inline-loader": "^0.8.0",
     "svgo": "^1.0.5",
     "svgo-loader": "^2.1.0",
     "testcafe": "^0.20.5",
